+++ docs/_pages/releases.md
@@ -12,7 +12,7 @@ sidebar:
 ### What's New
 
 ### Fixes
-* `ContainItemsAssignableTo` Now Expects At Least One Item Assignable to `T` - [#1765](https://github.com/fluentassertions/fluentassertions/pull/1765)
+* `ContainItemsAssignableTo` now expects at least one item assignable to `T` - [#1765](https://github.com/fluentassertions/fluentassertions/pull/1765)
 
 ## 6.3.0